+++ Sources/Sentry/include/SentryMetricProfiler.h
@@ -7,6 +7,7 @@
 @class SentryNSProcessInfoWrapper;
 @class SentryNSTimerWrapper;
 @class SentrySystemWrapper;
+@class SentryTransaction;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -16,16 +17,34 @@ SENTRY_EXTERN NSString *const kSentryMetricProfilerSerializationKeyCPUUsageForma
 SENTRY_EXTERN NSString *const kSentryMetricProfilerSerializationUnitBytes;
 SENTRY_EXTERN NSString *const kSentryMetricProfilerSerializationUnitPercentage;
 
+// The next two types are technically the same as far as the type system is concerned, but they
+// actually contain different mixes of value types, so define them separately. If they ever change,
+// the usage sites already specify which type each should be.
+
+/**
+ * A structure to hold a single metric reading and the time it was taken, as a dictionary with keyed
+ * values either of type NSNumber for the reading value, or NSString for the timestamp (we just
+ * encode @cuint64\_t as a string since JSON doesn't officially support it).
+ */
+typedef NSDictionary<NSString *, id /* <NSNumber, NSString> */> SentrySerializedMetricReading;
+
+/**
+ * A structure containing the timeseries of values for a particular metric type, as a dictionary
+ * with keyed values either of type NSString, for unit names, or an array of metrics entries
+ * containing the values and timestamps in the above typedef.
+ */
+typedef NSDictionary<NSString *, id /* <NSString, NSArray<SentrySerializedMetricEntry *>> */>
+    SentrySerializedMetricEntry;
+
 /**
  * A profiler that gathers various time-series and event-based metrics on the app process, such as
  * CPU and memory usage timeseries and thermal and memory pressure warning notifications.
  */
 @interface SentryMetricProfiler : NSObject
 
-- (instancetype)initWithProfileStartTime:(uint64_t)profileStartTime
-                      processInfoWrapper:(SentryNSProcessInfoWrapper *)processInfoWrapper
-                           systemWrapper:(SentrySystemWrapper *)systemWrapper
-                            timerWrapper:(SentryNSTimerWrapper *)timerWrapper;
+- (instancetype)initWithProcessInfoWrapper:(SentryNSProcessInfoWrapper *)processInfoWrapper
+                             systemWrapper:(SentrySystemWrapper *)systemWrapper
+                              timerWrapper:(SentryNSTimerWrapper *)timerWrapper;
 - (void)start;
 - (void)stop;
 
@@ -43,7 +62,8 @@ SENTRY_EXTERN NSString *const kSentryMetricProfilerSerializationUnitPercentage;
  * }
  * @endcode
  */
-- (NSMutableDictionary<NSString *, id> *)serialize;
+- (NSMutableDictionary<NSString *, SentrySerializedMetricEntry *> *)serializeForTransaction:
+    (SentryTransaction *)transaction;
 
 @end