+++ Tests/SentryTests/SentryClientTests.swift
@@ -1,4 +1,5 @@
 import Sentry
+import SentryTestUtils
 import XCTest
 
 // swiftlint:disable file_length