+++ Tests/SentryTests/SentryCrash/SentryUIDeviceWrapperTests.swift
@@ -1,3 +1,4 @@
+import SentryTestUtils
 import XCTest
 
 #if os(iOS)