+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixes 
+
+- Stop using UIScreen.main (#2762)
+
 ## 8.3.0
 
 ### Fixes
@@ -9,7 +15,6 @@
 - InApp for MetricKit stack traces (#2739)
 - Mutating while enumerating crash in Tracer (#2744)
 - Normalize profiling timestamps relative to transaction start (#2729)
-- Stop using UIScreen.main (#2762)
 
 ## 8.2.0